Decompile htTick

+++ src/include/game/data/data_02da90.h
@@ -39,7 +39,7 @@ extern u32 g_CiDeviceStageFlags[10];
 extern u32 device_descriptions[];
 extern u32 device_tips1[];
 extern u32 device_tips2[];
-extern u32 var80088bb8;
+extern u8 var80088bb8;
 extern u32 var80088bbc;
 extern u32 var80088bc8;
 extern u32 var80088be4;
